requirements.txt
matplotlib
sv-ttk
Pillow
tkcalendar
pyinstaller